+++ Photo Club Hub/View/MemberPortfoliosView/MemberPortfolioListView2626.swift
@@ -99,8 +99,11 @@ struct MemberPortfolioListView2626: View {
             ToolbarItemGroup(placement: toolbarItemPlacement) {
 
                 NavigationLink(destination: {
-                    let predicateAll = NSPredicate(format: "TRUEPREDICATE")
-                    OrganizationListView2626(predicate: predicateAll)
+                    let predicate: NSPredicate = Settings.showTestClubs ?
+                        NSPredicate(format: "TRUEPREDICATE") : // all clubs
+                        NSPredicate(format: "nickName_ CONTAINS %@",
+                                    argumentArray: ["fg"]) // TODO assumes pred passed to OrganizationListView2626 works
+                    OrganizationListView2626(predicate: predicate)
                 }, label: {
                     Image("mappin.ellipse.rectangle")
                         .font(.title)